[MIG] account_fiscal_position_vat_check: Migration to 10.0
Rename manifest files
