+++ account_fiscal_position_vat_check/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import account_invoice
+from . import partner
