--- account_fiscal_position_vat_check/partner_view.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!--
-    Copyright (C) 2013-2014 Akretion (http://www.akretion.com/)
-    @author Alexis de Lattre <alexis.delattre@akretion.com>
-    The licence is in the file __openerp__.py
--->
-
-<openerp>
-<data>
-
-
-<record id="view_partner_property_form" model="ir.ui.view">
-    <field name="name">customer.must.have.vat.fiscal_position_form</field>
-    <field name="model">res.partner</field>
-    <field name="inherit_id" ref="account.view_partner_property_form" />
-    <field name="arch" type="xml">
-        <field name="property_account_position" position="attributes">
-            <attribute name="on_change">fiscal_position_change(property_account_position, vat, customer)</attribute>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</openerp>
